+++ dangerfile.ts
@@ -115,10 +115,4 @@ if (!isBot) {
     );
   });
   raiseIssueAboutPaths(fail, onlyTestFiles, 'an `only` was left in the test');
-
-  // Politely ask for their name in the authors file
-  message('Please add your name and email to the AUTHORS file (optional)');
-  message(
-    'If this was a change that affects the external API, please update the docs and post a link to the PR in the discussion',
-  );
 }