+++ packages/apollo-link-retry/src/retryLink.ts
@@ -133,6 +133,7 @@ class RetryableOperation<TValue = any> {
   private onNext = (value: any) => {
     this.values.push(value);
     for (const observer of this.observers) {
+      if (!observer) continue;
       observer.next(value);
     }
   };
@@ -140,6 +141,7 @@ class RetryableOperation<TValue = any> {
   private onComplete = () => {
     this.complete = true;
     for (const observer of this.observers) {
+      if (!observer) continue;
       observer.complete();
     }
   };
@@ -148,16 +150,19 @@ class RetryableOperation<TValue = any> {
     this.retryCount += 1;
 
     // Should we retry?
-    const shouldRetry = await this.retryIf(this.retryCount, this.operation, error)
+    const shouldRetry = await this.retryIf(
+      this.retryCount,
+      this.operation,
+      error,
+    );
     if (shouldRetry) {
-      this.scheduleRetry(
-        this.delayFor(this.retryCount, this.operation, error),
-      );
+      this.scheduleRetry(this.delayFor(this.retryCount, this.operation, error));
       return;
     }
 
     this.error = error;
     for (const observer of this.observers) {
+      if (!observer) continue;
       observer.error(error);
     }
   };